fix(Decisions): clear organisation
rename templates using simple numbering to reflect the "inverted pyramid"
instead of UIDs, and add a header to each template to clarify status and
clearly associate each document with a PR and sponsor.

There is no need to add a "contributors" and "updated" section to the
header, as Github (the PR) is already doing that.